DRY sample/dataset constraints code
